Optimize block validation
Don't recalculate the hash of the previous block when validating a
block and use the recorded hash instead. This is safe only if the
provided previous block is a valid block, and that is true in the two
scenarios where a block is validated: when adding it to the chain and
when validating the entire chain. When adding a block to the chain, that
block is validated so it is ensured that only valid blocks make it into
the chain. When validating the entire chain, blocks are validated from
the oldest to the newest so the previous block on each validation was
already validated in the previous iteration.

Fixes #16.